Fix incorrect logic for ending elements (#55)
We used to use the last element to end normal closing elements. But in retrospect that approach was obviously flawed, because a sequence of group-open, path-open, path-close, group-close would have caused it to pass the wrong element to the parser objects (path instead of group).

The original logic for this was pretty messed up, and one of the consequences was that saving the last element instead of using a "fresh" element from the xml parser actually fixed it for some of the drawables I was using at the time.

This pull request corrects this issue.