Allow Tint Color to be associated with a VectorDrawable (#65)
This is useful if a user is working with a predefined library of Drawables, such as our brand icons, and wishes to tint one in a different way before passing into a UI component that doesn't take a tint color with the Drawable. It doesn't make sense to add a new drawable only for the purposes of having a new color.

This also matches the API available in Android, though it's slightly complicated by the fact that in Android there's no such thing as a VectorDrawableView; The VectorDrawable is itself a View.